backend/requirements.txt
setuptools
fastapi
uvicorn
rclpy
pyyaml
